+++ projects/Mallard/src/components/front/items/helpers/sizes.ts
@@ -1,0 +1,35 @@
+import {
+    ItemSizes,
+    PageLayoutSizes,
+    getPageLayoutSizeXY,
+} from '../../helpers/helpers'
+
+export const getImageHeight = ({ story, layout }: ItemSizes) => {
+    if (layout === PageLayoutSizes.tablet) {
+        if (story.height >= 4) {
+            return '50%'
+        }
+        if (story.height >= 3) {
+            return '66.66%'
+        }
+        if (story.height >= 2) {
+            return '50%'
+        }
+        return '75.5%'
+    }
+    if (layout === PageLayoutSizes.mobile) {
+        if (story.height > 4) {
+            return '75.5%'
+        }
+        return '50%'
+    }
+}
+
+export const isSmallItem = (size: ItemSizes) => {
+    return size.story.width <= 1
+}
+
+export const isFullWidthItem = (size: ItemSizes) => {
+    const { width } = getPageLayoutSizeXY(size.layout)
+    return size.story.width >= width
+}
